LPS-54881 decouple query from index operations
