+++ portal-impl/src/com/liferay/portal/search/lucene/LuceneHelper.java
@@ -14,8 +14,6 @@
 
 package com.liferay.portal.search.lucene;
 
-import com.liferay.portal.kernel.search.BooleanClauseOccur;
-
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -26,7 +24,6 @@ import java.util.Set;
 import org.apache.lucene.analysis.Analyzer;
 import org.apache.lucene.document.Document;
 import org.apache.lucene.index.Term;
-import org.apache.lucene.search.BooleanQuery;
 import org.apache.lucene.search.IndexSearcher;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.highlight.Formatter;
@@ -43,102 +40,6 @@ public interface LuceneHelper {
 
 	public void addDocument(long companyId, Document document)
 		throws IOException;
-
-	public void addExactTerm(
-		BooleanQuery booleanQuery, String field, boolean value);
-
-	public void addExactTerm(
-		BooleanQuery booleanQuery, String field, double value);
-
-	public void addExactTerm(
-		BooleanQuery booleanQuery, String field, int value);
-
-	public void addExactTerm(
-		BooleanQuery booleanQuery, String field, long value);
-
-	public void addExactTerm(
-		BooleanQuery booleanQuery, String field, short value);
-
-	public void addExactTerm(
-		BooleanQuery booleanQuery, String field, String value);
-
-	public void addNumericRangeTerm(
-		BooleanQuery booleanQuery, String field, Integer startValue,
-		Integer endValue);
-
-	public void addNumericRangeTerm(
-		BooleanQuery booleanQuery, String field, Long startValue,
-		Long endValue);
-
-	public void addNumericRangeTerm(
-		BooleanQuery booleanQuery, String field, short startValue,
-		short endValue);
-
-	public void addNumericRangeTerm(
-		BooleanQuery booleanQuery, String field, Short startValue,
-		Short endValue);
-
-	/**
-	 * @deprecated As of 6.2.0, replaced by {@link
-	 *             #addNumericRangeTerm(BooleanQuery, String, Long, Long)}
-	 */
-	@Deprecated
-	public void addNumericRangeTerm(
-		BooleanQuery booleanQuery, String field, String startValue,
-		String endValue);
-
-	public void addRangeTerm(
-		BooleanQuery booleanQuery, String field, int startValue, int endValue);
-
-	public void addRangeTerm(
-		BooleanQuery booleanQuery, String field, long startValue,
-		long endValue);
-
-	public void addRangeTerm(
-		BooleanQuery booleanQuery, String field, Long startValue,
-		Long endValue);
-
-	public void addRangeTerm(
-		BooleanQuery booleanQuery, String field, String startValue,
-		String endValue);
-
-	public void addRequiredTerm(
-		BooleanQuery booleanQuery, String field, boolean value);
-
-	public void addRequiredTerm(
-		BooleanQuery booleanQuery, String field, double value);
-
-	public void addRequiredTerm(
-		BooleanQuery booleanQuery, String field, int value);
-
-	public void addRequiredTerm(
-		BooleanQuery booleanQuery, String field, long value);
-
-	public void addRequiredTerm(
-		BooleanQuery booleanQuery, String field, short value);
-
-	public void addRequiredTerm(
-		BooleanQuery booleanQuery, String field, String value);
-
-	public void addRequiredTerm(
-		BooleanQuery booleanQuery, String field, String value, boolean like);
-
-	public void addRequiredTerm(
-		BooleanQuery booleanQuery, String field, String[] values, boolean like);
-
-	public void addTerm(BooleanQuery booleanQuery, String field, long value);
-
-	public void addTerm(BooleanQuery booleanQuery, String field, String value);
-
-	public void addTerm(
-		BooleanQuery booleanQuery, String field, String value, boolean like);
-
-	public void addTerm(
-		BooleanQuery booleanQuery, String field, String value, boolean like,
-		BooleanClauseOccur booleanClauseOccur);
-
-	public void addTerm(
-		BooleanQuery booleanQuery, String field, String[] values, boolean like);
 
 	/**
 	 * @deprecated As of 7.0.0, replaced by {@link #releaseIndexSearcher(long,
